pyproject.toml
[tool.poetry]
name = "keycloak-demo"
version = "0.1.0"
description = "test flask app for keycloak"
authors = ["wa wa <wa@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
python-keycloak = "^3.7.0"
flask = "^3.0.2"
flask-sqlalchemy = "^3.1.1"
psycopg2-binary = "^2.9.9"

[tool.poetry.dev-dependencies]
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
